Configures users with local filesystem
Modifies the user creation script to provision users
with a local filesystem home directory instead of relying
on S3. Removes S3-related configuration and updates
the script to use `/data/{username}` as the home directory
since s3 is mounted under /data in the container.

Fixes an oversight that gave users access to the container root fs 🤦