Adds DNS zones
Creates public and private hosted zones in Route 53 to map instances to their IP address.

This allows access to OpenVPN and the SFTPGo UI using a domain name within the VPC, improving usability and security by avoiding the need to use the private IP directly.

Also, increase default TTL for DNS records to 900.

To resolve the private record the client DNS will have to be configured to lookup from AWS VPC DNS = 169.254.169.253